Moved TimeProvider class into library utils/ directory.
